requirements.txt
attrs==24.2.0
fastapi==0.115.6
langchain==0.3.4
langchain-anthropic==0.2.3
langchain-google-genai==2.0.1
langchain-openai==0.2.3
openai==1.52.2
pdf2image==1.17.0
pillow==10.4.0
pydantic==2.9.2
python-dotenv==1.0.1
python-magic==0.4.27
python-multipart==0.0.20